Move OrderConstantsByVisibilityRector to Class_
